Relocate some functions to correct directory
